set up passenger service router
