+++ server/index.js
@@ -9,7 +9,11 @@ const bodyParser = require('body-parser');
 const { DBHelper } = require('./shared/DB');
 const {Pagination} = require('./shared/pagination')
 const {PORT} = require("./config/dev");
+
+// Controllers 
 const { DriverController } = require('./internal/driver/controller/controller');
+const { PassengerController } = require('./internal/passenger/controller/controller');
+// const { AdminController } = require('./internal/admin/controller/controller');
 
 // Repositories 
 const { DriverRepository } = require('./internal/driver/repository/repository');
@@ -17,6 +21,7 @@ const { PassengerRepository } = require('./internal/passenger/repository/reposit
 const { TripRepository } = require('./internal/trip/repository/repository');
 const { PaymentRepository } = require('./internal/payment/repository/repository');
 const { getDriverRouter } = require('./internal/driver');
+const { getPassengerRouter } = require('./internal/passenger');
 
 
 // App Config  
@@ -37,16 +42,19 @@ app.use(bodyParser.urlencoded({extended: true}));
 const dbHelper = new DBHelper(mysql, config, new Pagination())
 connection = dbHelper.createConnection()
 const driverRepo = new DriverRepository(dbHelper)
+const passengerRepo = new PassengerRepository(dbHelper)
 const tripRepo = new TripRepository(dbHelper)
 const paymentRepo = new PaymentRepository(dbHelper)
-const driver = new DriverController(driverRepo, tripRepo, paymentRepo)
 
+const driver = new DriverController(driverRepo, tripRepo, paymentRepo)
+const passenger = new PassengerController(passengerRepo, tripRepo, paymentRepo)
 
 // Routes 
 app.use('/api/driver',getDriverRouter(driver))
+app.use('/api/passenger', getPassengerRouter(passenger))
+
 
 // Server Config 
 app.listen(9000, () => {
     console.log("Running the backend server on port 9000");
-
 })
