+++ server/internal/passenger/controller/controller.js
@@ -1,0 +1,9 @@
+class PassengerController {
+    constructor(passengerRepo, tripRepo, paymentRepo) {
+        this._passengerRepo = passengerRepo
+        this._tripRepo = tripRepo
+        this._paymentRepo = paymentRepo
+    }
+}
+
+module.exports = {PassengerController}
